\name{bsseq-defunct}
\alias{read.modkit}
\alias{read.modbam2bed}
\title{Defunct functions in package \sQuote{bsseq}}
\description{These functions are defunct and no longer available.}

\details{
  Defunct functions are: \code{read.modkit},
  \code{read.modbam2bed}. Both functions are replaced by
  \code{read.bedMethyl}.
}
