\name{bsseq-deprecated}
\alias{bsseq-deprecated}
\title{Deprecated functions in package \sQuote{bsseq}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{bsseq} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{read.modbam2bed: \code{\link{read.bedMethyl}}}
    \item{read.modkit: \code{\link{read.bedMethyl}}}

  }
}
